fix uniswap v4 decoding when fees are paid via internal transactions
